+++ pkg/scaling/executor/scale_executor.go
@@ -1,19 +1,34 @@
+/*
+Copyright 2021 The KEDA Authors
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
 package executor
 
 import (
 	"context"
 	"fmt"
 
-	"k8s.io/client-go/tools/record"
-
 	"github.com/go-logr/logr"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/client-go/scale"
-	"sigs.k8s.io/controller-runtime/pkg/client"
+	"k8s.io/client-go/tools/record"
+	runtimeclient "sigs.k8s.io/controller-runtime/pkg/client"
 	logf "sigs.k8s.io/controller-runtime/pkg/log"
 
-	kedav1alpha1 "github.com/kedacore/keda/v2/api/v1alpha1"
+	kedav1alpha1 "github.com/kedacore/keda/v2/apis/keda/v1alpha1"
 )
 
 const (
@@ -28,7 +43,7 @@ type ScaleExecutor interface {
 }
 
 type scaleExecutor struct {
-	client           client.Client
+	client           runtimeclient.Client
 	scaleClient      scale.ScalesGetter
 	reconcilerScheme *runtime.Scheme
 	logger           logr.Logger
@@ -36,7 +51,7 @@ type scaleExecutor struct {
 }
 
 // NewScaleExecutor creates a ScaleExecutor object
-func NewScaleExecutor(client client.Client, scaleClient scale.ScalesGetter, reconcilerScheme *runtime.Scheme, recorder record.EventRecorder) ScaleExecutor {
+func NewScaleExecutor(client runtimeclient.Client, scaleClient scale.ScalesGetter, reconcilerScheme *runtime.Scheme, recorder record.EventRecorder) ScaleExecutor {
 	return &scaleExecutor{
 		client:           client,
 		scaleClient:      scaleClient,
@@ -47,16 +62,16 @@ func NewScaleExecutor(client client.Client, scaleClient scale.ScalesGetter, reco
 }
 
 func (e *scaleExecutor) updateLastActiveTime(ctx context.Context, logger logr.Logger, object interface{}) error {
-	var patch client.Patch
+	var patch runtimeclient.Patch
 
 	now := metav1.Now()
-	runtimeObj := object.(runtime.Object)
+	runtimeObj := object.(runtimeclient.Object)
 	switch obj := runtimeObj.(type) {
 	case *kedav1alpha1.ScaledObject:
-		patch = client.MergeFrom(obj.DeepCopy())
+		patch = runtimeclient.MergeFrom(obj.DeepCopy())
 		obj.Status.LastActiveTime = &now
 	case *kedav1alpha1.ScaledJob:
-		patch = client.MergeFrom(obj.DeepCopy())
+		patch = runtimeclient.MergeFrom(obj.DeepCopy())
 		obj.Status.LastActiveTime = &now
 	default:
 		err := fmt.Errorf("unknown scalable object type %v", obj)
@@ -72,15 +87,15 @@ func (e *scaleExecutor) updateLastActiveTime(ctx context.Context, logger logr.Lo
 }
 
 func (e *scaleExecutor) setCondition(ctx context.Context, logger logr.Logger, object interface{}, status metav1.ConditionStatus, reason string, message string, setCondition func(kedav1alpha1.Conditions, metav1.ConditionStatus, string, string)) error {
-	var patch client.Patch
+	var patch runtimeclient.Patch
 
-	runtimeObj := object.(runtime.Object)
+	runtimeObj := object.(runtimeclient.Object)
 	switch obj := runtimeObj.(type) {
 	case *kedav1alpha1.ScaledObject:
-		patch = client.MergeFrom(obj.DeepCopy())
+		patch = runtimeclient.MergeFrom(obj.DeepCopy())
 		setCondition(obj.Status.Conditions, status, reason, message)
 	case *kedav1alpha1.ScaledJob:
-		patch = client.MergeFrom(obj.DeepCopy())
+		patch = runtimeclient.MergeFrom(obj.DeepCopy())
 		setCondition(obj.Status.Conditions, status, reason, message)
 	default:
 		err := fmt.Errorf("unknown scalable object type %v", obj)